chore: migrate listbox test
